requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
charset-normalizer==3.1.0
discord.py==2.2.2
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
motor==3.1.2
multidict==6.0.4
pycodestyle==2.10.0
pymongo==4.3.3
yarl==1.9.1
